+++ src/routes/(public)/page-sitemap.xml/+server.ts
@@ -1,0 +1,66 @@
+import type { RequestHandler } from './$types';
+import { packages } from '$lib/data/packages';
+
+export const GET: RequestHandler = async () => {
+	const staticPages = [
+		'',
+		'about-us',
+		'contact-us',
+		'faq',
+		'privacy-policy',
+		'terms-of-service',
+		'gdpr',
+		'cookie-policy',
+		'meet-the-team',
+		'sitemap',
+		'services',
+		'packages',
+		'blog'
+	];
+
+	// Base URL of the website
+	const baseUrl = 'https://malagaeventgear.com';
+
+	const pagesXml = staticPages
+		.map((page) => {
+			const loc = `${baseUrl}/${page}${page ? '/' : ''}`;
+			return `	<url>
+		<loc>${loc}</loc>
+		<lastmod>${new Date().toISOString().split('T')[0]}T00:00:00+00:00</lastmod>
+	</url>`;
+		})
+		.join('\n');
+
+	const packagesXml = packages
+		.map((pkg) => {
+			const loc = `${baseUrl}${pkg.route}`;
+			const imageBlock = pkg.image
+				? `\n		<image:image>
+			<image:loc>${pkg.image}</image:loc>
+		</image:image>`
+				: '';
+
+			return `	<url>
+		<loc>${loc}</loc>
+		<lastmod>${new Date().toISOString().split('T')[0]}T00:00:00+00:00</lastmod>${imageBlock}
+	</url>`;
+		})
+		.join('\n');
+
+	const xml = `<?xml version="1.0" encoding="UTF-8"?>
+<urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9"
+        xmlns:image="http://www.google.com/schemas/sitemap-image/1.1"
+        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+        xsi:schemaLocation="http://www.sitemaps.org/schemas/sitemap/0.9 http://www.sitemaps.org/schemas/sitemap/0.9/sitemap.xsd http://www.google.com/schemas/sitemap-image/1.1 http://www.google.com/schemas/sitemap-image/1.1/sitemap-image.xsd">
+${pagesXml}
+${packagesXml}
+</urlset>`;
+
+	return new Response(xml, {
+		headers: {
+			'Content-Type': 'application/xml; charset=utf-8',
+			'Cache-Control': 'public, max-age=3600, s-maxage=86400',
+			'X-Content-Type-Options': 'nosniff'
+		}
+	});
+};