+++ src/routes/(public)/post-sitemap.xml/+server.ts
@@ -1,0 +1,16 @@
+import type { RequestHandler } from './$types';
+
+export const GET: RequestHandler = async () => {
+	const xml = `<?xml version="1.0" encoding="UTF-8"?>
+<urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9"
+        xmlns:image="http://www.google.com/schemas/sitemap-image/1.1">
+</urlset>`;
+
+	return new Response(xml, {
+		headers: {
+			'Content-Type': 'application/xml; charset=utf-8',
+			'Cache-Control': 'public, max-age=3600, s-maxage=86400',
+			'X-Content-Type-Options': 'nosniff'
+		}
+	});
+};